LayoutTests/accessibility/corresponding-control-deleted-crash-expected.txt

CONSOLE MESSAGE: line 26: TypeError: 'undefined' is not a function (evaluating 'label.webkitCreateShadowRoot()')
Make sure that a debug assert is not triggered when a call to RenderBlock::deleteLineBoxTree calls AccessibilityRenderObject::accessibilityIsIgnored which may require the AXObject for a node that is being deleted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
